правка переменных

+++ work1/script.php
@@ -1,4 +1,4 @@
 <?php 
 $user = readline("Ваше имя?") ;
-$age = readline("Сколько вам лет?");
+$age = (int)readline("Сколько вам лет?");
 echo "Вас зовут $user , вам $age лет.";
